requirements.txt
openai-whisper
openwakeword
pyttsx3
soundfile
pyaudio
requests
